Split up resolver tests
